requirements.txt
openai==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
requests==2.31.0
tenacity==8.2.3
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==7.0.0
flask==3.0.0
